# perf(token): Speed up 'take_until' for str/char 