Before/after (with simd)
```
find_slice/byte/empty   time:   [8.6164 ns 9.0722 ns 9.5994 ns]
                        change: [-43.031% -39.731% -35.814%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/slice/empty  time:   [8.9148 ns 9.4395 ns 9.9990 ns]
                        change: [-40.889% -36.871% -32.492%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/byte/start   time:   [349.03 ns 382.08 ns 418.18 ns]
                        change: [-66.578% -63.083% -59.659%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/slice/start  time:   [621.33 ns 662.95 ns 707.40 ns]
                        change: [-55.891% -51.971% -47.614%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/byte/medium  time:   [535.01 ns 560.18 ns 586.97 ns]
                        change: [-68.347% -65.802% -62.864%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/slice/medium time:   [681.00 ns 725.15 ns 772.02 ns]
                        change: [-58.167% -54.656% -50.939%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/byte/large   time:   [12.233 µs 12.845 µs 13.494 µs]
                        change: [-97.200% -96.959% -96.698%] (p = 0.00 < 0.05)
                        Performance has improved.
find_slice/slice/large  time:   [12.739 µs 13.314 µs 13.912 µs]
                        change: [-91.483% -90.773% -90.011%] (p = 0.00 < 0.05)
                        Performance has improved.
```